requirements.txt
Django==4.1.5
Jinja2==3.1.2
python-decouple==3.7
requests==2.28.1
